Implement optimized serialization for item states and introduce encoding for item types, groups, interactable texts, and wave states
- Added efficient serialization methods for ItemState, reducing data size during network transmission.
- Introduced item type encoding to map item strings to unique numeric IDs for improved serialization efficiency.
- Implemented group encoding to represent group types as uint8 IDs, enhancing performance in network communications.
- Added interactable text encoding to streamline the representation of interactable display names.
- Introduced wave state encoding to efficiently manage wave states using sequential IDs.
- Updated relevant extensions and events to utilize the new encoding and serialization methods, ensuring consistency across the codebase.

+++ packages/game-client/src/extensions/poison.ts
@@ -32,8 +32,9 @@ export class ClientPoison extends BaseClientExtension {
   }
 
   public deserializeFromBuffer(reader: BufferReader): this {
-    this.maxDamage = reader.readFloat64();
-    this.totalDamage = reader.readFloat64();
+    // maxDamage and totalDamage are UInt8 (0-255)
+    this.maxDamage = reader.readUInt8();
+    this.totalDamage = reader.readUInt8();
     return this;
   }
 }
